sync

+++ models/product.go
@@ -7,5 +7,5 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 type Product struct {
 	ID    primitive.ObjectID `bson:"_id"`
 	Name  string             `bson:"name"`
-	Price int                `bson:"price"`
+	Price float64            `bson:"price"`
 }
